Retrabajo Product V1.5

+++ com.inn.almacen/src/main/java/com/inn/almacen/POJO/Supplier.java
@@ -27,7 +27,7 @@ public class Supplier implements Serializable {
     private String razonSocial;
 
     @Column(name = "ruc")
-    private Integer ruc;
+    private Long ruc;
 
     @Column(name = "contacto")
     private Integer contacto;
